Moved the shared methods of outcome and external mccfr into the super class _MCCFRSolverBase. Changed init constant of _lookup_infostate_info in external mccfr to 1/1e6
