+++ open_spiel/python/algorithms/cfr.py
@@ -114,7 +114,7 @@ def _update_average_policy(average_policy, info_state_nodes):
 
 
 class _CFRSolverBase(object):
-  r"""A base classe for both CFR and CFR-BR.
+  r"""A base class for both CFR and CFR-BR.
 
   The main iteration loop is implemented in `evaluate_and_update_policy`:
